+++ src/lib.rs
@@ -8,20 +8,13 @@ use mouse_rs::Mouse;
 use mouse_rs::types::Point;
 use spinners::{Spinner, Spinners};
 
-/// Empty enum to emulate the never type `!` and avoid using `#![feature(never_type)]`.
-///
-/// For additional information, see:
-/// * <https://rust-reference.irust.net/en-us/types/never.html>
-/// * <https://github.com/rust-lang/rust/issues/35121>
-pub enum Void {}
-
 /// Check the mouse position every `interval`; jiggle the mouse pointer and scroll the wheel if the
 /// position hasn't changed.
 ///
 /// ## Errors
 ///
 /// Returns a generic error in case something goes wrong.
-pub fn run(interval: Duration) -> Result<Void, Box<dyn std::error::Error>> {
+pub fn run(interval: Duration) -> Result<(), Box<dyn std::error::Error>> {
     let mouse = Mouse::new();
     let mut old_position = mouse.get_position()?;
     let is_same_pos = |p: &Point, q: &Point| p.x == q.x && p.y == q.y;